Move linux-only getProxyMode tests to a linux-only file.
